spike(okf): close Astra P2 #5b, fractional wall-aligned default clock
The default trusted_clock computed int(now) + int(elapsed), so an entry at
wall 1299.75 (now 1299) whose read ended at 1300.25 still read 1299 and
released after the absolute deadline 1300. The default clock is now
fractional and wall-aligned: it anchors now to the whole second of entry
and adds the exact wall time elapsed since that boundary, never running
backwards; Registry.try_consume compares the fractional sample without an
integer cast. Injected clocks are unchanged.

Regressions on the production clock path (time.time patched, no injected
clock): deadline crossed during result read / source probe / output
probe / render, and through the actual CLI helper -> REJECTED
request_expired, no value, nonce intact; a fractional just-inside case
still releases. Hermetic 133 pass; live suite 7/7 and the seven live CLI
cases rerun; cases.json regenerated at this head.

+++ examples/okf_attested_computation/receipt_store.py
@@ -161,8 +161,8 @@ class Registry:
       try:
         self._conn.execute("BEGIN IMMEDIATE")
         if deadline is not None:
-          sample = int(clock()) if clock is not None else int(time.time())
-          if sample >= deadline:
+          sample = float(clock()) if clock is not None else time.time()
+          if sample >= float(deadline):
             self._conn.execute("ROLLBACK")
             return "expired"
         cur = self._conn.execute(
